toyhana-back/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.31.0
httpx==0.28.1
pydantic==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.1
pyjwt==2.10.1
Pillow==11.0.0
pillow-heif==0.21.0
